The segmentation fault was due to cat, when the file does not end. I added the global function str2ull, which converts string to unsigned long long. This is needed for period_length_1ns.

+++ mesyread/kangaroo.cpp
@@ -144,7 +144,7 @@ void TKangaroo::init_histo() {
     zHisto->set_filename(filename);
     zHisto->set_first_left_edge(0.0);
     zHisto->set_last_right_edge(
-     (float)zPeriodeLength_1ns);
+     (double)zPeriodeLength_1ns);
     zHisto->place_bins();
     zHisto->set_flipper_off();
 }
